+++ packages/descriptions/src/index.tsx
@@ -3,13 +3,13 @@ import { Descriptions, Space, Form } from 'antd';
 import { EditOutlined, CloseOutlined, CheckOutlined } from '@ant-design/icons';
 import toArray from 'rc-util/lib/Children/toArray';
 import ProForm, { ProFormField } from '@ant-design/pro-form';
-import type { ProFieldFCMode } from '@ant-design/pro-field';
 import type {
   ProSchema,
   ProCoreActionType,
   RowEditableConfig,
   UseEditableMapUtilType,
   ProFieldValueType,
+  ProSchemaComponentTypes,
 } from '@ant-design/pro-utils';
 import {
   InlineErrorFormItem,
@@ -25,8 +25,9 @@ import type { DescriptionsItemProps } from 'antd/lib/descriptions/Item';
 import type { DescriptionsProps } from 'antd/lib/descriptions';
 import type { RequestData } from './useFetchData';
 import useFetchData from './useFetchData';
+import type { ProFieldFCMode } from '@ant-design/pro-utils';
 
-export type ProDescriptionsItemProps<T = {}> = ProSchema<
+export type ProDescriptionsItemProps<T = Record<string, any>, ValueType = 'text'> = ProSchema<
   T,
   Omit<DescriptionsItemProps, 'children'> & {
     // 隐藏这个字段，是个语法糖，方便一下权限的控制
@@ -36,11 +37,16 @@ export type ProDescriptionsItemProps<T = {}> = ProSchema<
     ellipsis?: boolean;
     mode?: ProFieldFCMode;
     children?: React.ReactNode;
-  }
+  },
+  ProSchemaComponentTypes,
+  ValueType
 >;
 export type ProDescriptionsActionType = ProCoreActionType;
 
-export type ProDescriptionsProps<RecordType = {}> = DescriptionsProps & {
+export type ProDescriptionsProps<
+  RecordType = Record<string, any>,
+  ValueType = 'text'
+> = DescriptionsProps & {
   /**
    * params 参数
    * params 改变的时候会触发 reload
@@ -55,7 +61,7 @@ export type ProDescriptionsProps<RecordType = {}> = DescriptionsProps & {
    */
   request?: (params: Record<string, any>) => Promise<RequestData>;
 
-  columns?: ProDescriptionsItemProps<RecordType>[];
+  columns?: ProDescriptionsItemProps<RecordType, ValueType>[];
 
   /**
    * 一些简单的操作
@@ -337,7 +343,9 @@ const ProDescriptionsItem: React.FC<ProDescriptionsItemProps> = (props) => {
   return <Descriptions.Item {...props}>{props.children}</Descriptions.Item>;
 };
 
-const ProDescriptions = <RecordType extends {}>(props: ProDescriptionsProps<RecordType>) => {
+const ProDescriptions = <RecordType extends Record<string, any>, ValueType = 'text'>(
+  props: ProDescriptionsProps<RecordType, ValueType>,
+) => {
   const {
     request,
     columns,