+++ packages/table/src/demos/columnsStateMap.tsx
@@ -41,6 +41,7 @@ const columns: ProColumns<TableListItem>[] = [
     dataIndex: 'status',
     initialValue: 'all',
     filters: true,
+    onFilter: true,
     valueType: 'select',
     valueEnum: {
       all: { text: '全部', status: 'Default' },
@@ -76,6 +77,7 @@ export default () => {
   const [columnsStateMap, setColumnsStateMap] = useState<Record<string, ColumnsState>>({
     name: {
       show: false,
+      order: 2,
     },
   });
   return (