+++ packages/table/src/demos/dynamic-settings.tsx
@@ -43,9 +43,10 @@ const columns: ProColumnType<DataType>[] = [
     title: 'Action',
     key: 'action',
     sorter: true,
+    valueType: 'option',
     render: () => [
-      <a>Delete</a>,
-      <a className="ant-dropdown-link">
+      <a key="delete">Delete</a>,
+      <a key="link" className="ant-dropdown-link">
         More actions <DownOutlined />
       </a>,
     ],
@@ -77,7 +78,7 @@ const initData = {
   hasData: true,
   tableLayout: undefined,
   search: true,
-  span: 8,
+  span: 24,
   toolBarRender: true,
   collapseRender: true,
   options: {
@@ -177,6 +178,10 @@ const DynamicSettings = () => {
             <ProFormSelect
               tooltip={`search={{span:8}}`}
               options={[
+                {
+                  label: '24',
+                  value: 24,
+                },
                 {
                   label: '12',
                   value: 12,